project/Worker/requirements.txt
redis
pillow
rembg
onnxruntime